Add player script & refactoring names
<For Better Undrestanding>
