+++ Assets/Scripts/Player.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 82d6daf7a5e247fc871c953b7e18c44c
+timeCreated: 1693653570
